// src/app/modules/staff/staff-information-system/model/class-teacher-assignment-model.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface ClassTeacherAssignmentModel { 
    /**
     * Gets or sets
     */
    classId?: number;
    /**
     * Gets or sets
     */
    sectionId?: number;
    /**
     * Gets or sets
     */
    schoolBoardId?: number;
    /**
     * Gets or sets
     */
    classTeacherId?: number;
}